# api-docs.json
{
    "openapi": "3.0.3",
    "info": {
      "version": "1.1.0",
      "title": " NodeBucket - OpenAPI Documentation",
      "description": "NodeBucket - OpenAPI Documentation"
    },
    "servers": [
      {
        "url": "http://localhost:3000/api"
      }
    ],
    "tags": [
      {
        "name": "Employees",
        "description": "Functions involving Employees"
      },
      {
        "name": "Tasks",
        "description": "Functions involving Tasks"
      }
    ],
    "paths": {
      "/employees/{empId}": {
        "get": {
          "tags": ["Employees"],
          "summary": "findEmployeeById",
          "description": "Find an employee by their unique Employee ID.",
          "parameters": [
            {
              "name": "empId",
              "in": "path",
              "required": true,
              "description": "Employee ID",
              "schema": {
                "type": "number"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "empId": {
                        "type": "number",
                        "example": 1008
                      },
                      "firstName": {
                        "type": "string",
                        "example": "Clarke"
                      },
                      "lastName": {
                        "type": "string",
                        "example": "Griffin"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/employees/{empId}/tasks": {
        "get": {
          "tags": ["Tasks"],
          "summary": "findAllTasks",
          "description": "Find all tasks for an employee by entering their employee ID",
          "parameters": [
            {
              "name": "empId",
              "in": "path",
              "required": true,
              "description": "Enter an employee ID",
              "schema": {
                "type": "number"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Employee found and tasks displayed",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "todo": {
                        "text": "string",
                        "example": "Test to see if the createTask API works."
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "post": {
          "tags": ["Tasks"],
          "summary": "createTask",
          "description": "API for creating a new task for an Employee",
          "parameters": [
            {
              "name": "empId",
              "in": "path",
              "required": true,
              "description": "Enter an employee ID",
              "schema": {
                "text": {
                  "type": "string"
                },
                "category": {
                  "categoryName": "string",
                  "backgroundColor": "string"
                }
              }
            }],
            "requestBody": {
              "description": "Task info",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": ["empId"],
                    "properties": {
                      "text": {
                        "type": "string"
                      },
                      "category": {
                        "categoryName": {
                          "type": "string"
                      },
                      "backgroundColor": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
            },
          "responses": {
            "201": {
              "description": "The task has been created",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "todo": {
                        "text": "string",
                        "example": "Test to see if the createTask API works."
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
        }
      }
    },
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"]
  }
}